% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\docType{class}
\name{LatexTableTemplate-class}
\alias{LatexTableTemplate-class}
\alias{LatexTableTemplate}
\title{LatexTableTemplate}
\description{
LatexTableTemplate
}
\details{
Slot *_fn accept data as its input, the output should have length 0 (indicate this is a no-op) or length(data).

The used column_names should have the same length that defined in columns
}
\section{Slots}{

\describe{
\item{\code{caption}}{Default caption, contains 2 elements, caption\link{1} is the optional caption, and caption\link{2} is the optional label.}

\item{\code{columns}}{contains 2 elements, controlling the table format for name columns and data columns.}

\item{\code{column_names}}{Default column_names, could be empty (in this case, use data colnames instead).}

\item{\code{roundings}}{control the roundings of the column, its length could be 1 (global rounding) or length(column_names) (column rounding). Data displayed in percentage format has 2 less rounding digits.}

\item{\code{table_rules}}{define extra latex rules before the \\begin\{tabular\} command}

\item{\code{top_rules}}{define extra latex rules after the \\begin\{tabular\} command}

\item{\code{bottom_rules}}{define extra latex rules before the \\end\{tabular\} command}

\item{\code{footnotes}}{define extra latex rules after the \\end\{tabular\} command}

\item{\code{bold_fn}}{function control which item is wrapped with \\bold\{..\}}

\item{\code{italic_fn}}{function control which item is wrapped with \\italic\{..\}}

\item{\code{stars_fn}}{function calculate how much stars is used.}
}}

\section{Methods}{

\describe{
\item{\code{data(data, cap = character(0), row_name = character(0))}}{display the latex format with the given data, use caption \code{cap} and row_name \code{row_name} if provided.}
}}

\examples{
template <- LatexTableTemplate(caption=c("cap","lab"), columns = c("lll|","rrr\%"), column_names = c("1","2","3"), roundings=4L, table_rules = character(0), top_rules="\\\\toprules", bottom_rules = "\\\\bottom_rules", footnotes = character(0), bold_fn=function(x)NULL, italic_fn=function(x)NULL, stars_fn=function(x)NULL)
template$data(matrix(1:6/6,2,3))
# That will directly print:
# \begin{table}
#     \caption{cap\label{lab}}
#     \begin{tabular}{lll|rrr}
#         \toprules
#          &  &  & 1 & 2 & 3\\
#           &&& 0.1667 & 0.5000 & 83.33\%\\
#           &&& 0.3333 & 0.6667 & 100.00\%\\
#         \bottom_rules
#     \end{tabular}
# \end{table}
}
